refactor: move MainScreen into separate dir
To add module.scss file later.
