+++ packages/frontend/src/components/screens/MainScreen/MainScreen.tsx
@@ -7,35 +7,37 @@ import React, {
 } from 'react'
 import { C } from '@deltachat/jsonrpc-client'
 
-import Gallery from '../Gallery'
-import { useThreeDotMenu } from '../ThreeDotMenu'
-import ChatList from '../chat/ChatList'
-import { Avatar } from '../Avatar'
-import { Navbar, NavbarGroupLeft, NavbarGroupRight } from '../Navbar'
-import ConnectivityToast from '../ConnectivityToast'
-import MailingListProfile from '../dialogs/MessageListProfile'
-import SettingsStoreInstance, { useSettingsStore } from '../../stores/settings'
-import { Type } from '../../backend-com'
-import { InlineVerifiedIcon } from '../VerifiedIcon'
-import Button from '../Button'
-import Icon from '../Icon'
-import SearchInput from '../SearchInput'
-import MessageListView from '../MessageListView'
-import useChat from '../../hooks/chat/useChat'
-import useDialog from '../../hooks/dialog/useDialog'
-import useKeyBindingAction from '../../hooks/useKeyBindingAction'
-import useOpenViewGroupDialog from '../../hooks/dialog/useOpenViewGroupDialog'
-import useOpenViewProfileDialog from '../../hooks/dialog/useOpenViewProfileDialog'
-import useSelectLastChat from '../../hooks/chat/useSelectLastChat'
-import useTranslationFunction from '../../hooks/useTranslationFunction'
-import { KeybindAction } from '../../keybindings'
-import { selectedAccountId } from '../../ScreenController'
-import { openMapWebxdc } from '../../system-integration/webxdc'
-import { ChatView } from '../../contexts/ChatContext'
-import { ScreenContext } from '../../contexts/ScreenContext'
+import Gallery from '../../Gallery'
+import { useThreeDotMenu } from '../../ThreeDotMenu'
+import ChatList from '../../chat/ChatList'
+import { Avatar } from '../../Avatar'
+import { Navbar, NavbarGroupLeft, NavbarGroupRight } from '../../Navbar'
+import ConnectivityToast from '../../ConnectivityToast'
+import MailingListProfile from '../../dialogs/MessageListProfile'
+import SettingsStoreInstance, {
+  useSettingsStore,
+} from '../../../stores/settings'
+import { Type } from '../../../backend-com'
+import { InlineVerifiedIcon } from '../../VerifiedIcon'
+import Button from '../../Button'
+import Icon from '../../Icon'
+import SearchInput from '../../SearchInput'
+import MessageListView from '../../MessageListView'
+import useChat from '../../../hooks/chat/useChat'
+import useDialog from '../../../hooks/dialog/useDialog'
+import useKeyBindingAction from '../../../hooks/useKeyBindingAction'
+import useOpenViewGroupDialog from '../../../hooks/dialog/useOpenViewGroupDialog'
+import useOpenViewProfileDialog from '../../../hooks/dialog/useOpenViewProfileDialog'
+import useSelectLastChat from '../../../hooks/chat/useSelectLastChat'
+import useTranslationFunction from '../../../hooks/useTranslationFunction'
+import { KeybindAction } from '../../../keybindings'
+import { selectedAccountId } from '../../../ScreenController'
+import { openMapWebxdc } from '../../../system-integration/webxdc'
+import { ChatView } from '../../../contexts/ChatContext'
+import { ScreenContext } from '../../../contexts/ScreenContext'
 
 import type { T } from '@deltachat/jsonrpc-client'
-import CreateChat from '../dialogs/CreateChat'
+import CreateChat from '../../dialogs/CreateChat'
 
 type Props = {
   accountId?: number
